# Parallelise the CPU kernels for the conv ops.