--- a/Cargo.toml
+++ b/Cargo.toml
@@ -43,6 +43,7 @@ num_cpus = "1.15.0"
 num-traits = "0.2.15"
 rand = "0.8.5"
 rand_distr = "0.4.3"
+rayon = "1.7.0"
 safetensors = "0.3.1"
 serde = { version = "1.0.171", features = ["derive"] }
 serde_json = "1.0.99"