--- a/candle-core/Cargo.toml
+++ b/candle-core/Cargo.toml
@@ -24,6 +24,7 @@ num-traits = { workspace = true }
 num_cpus = { workspace = true }
 rand = { workspace = true }
 rand_distr = { workspace = true }
+rayon = { workspace = true }
 safetensors = { workspace = true }
 thiserror = { workspace = true }
 zip = { workspace = true }