--- a/candle-core/src/cpu_backend.rs
+++ b/candle-core/src/cpu_backend.rs
@@ -1032,7 +1032,7 @@ impl<'a> Map2 for Conv1D<'a> {
         let l_out = p.l_out();
         let dst_elems = p.c_out * l_out * p.b_size;
         // The output shape is [b_size, c_out, l_out]
-        let mut dst = vec![T::zero(); dst_elems];
+        let dst = vec![T::zero(); dst_elems];
 
         // TODO: Avoid making this copy if `inp` already has the appropriate layout.
         let mut inp_cont = vec![T::zero(); p.b_size * p.c_in * p.l_in];
@@ -1045,8 +1045,10 @@ impl<'a> Map2 for Conv1D<'a> {
             }
         }
 
+        let num_threads = crate::utils::get_num_threads();
+
         for offset in 0..p.k_size {
-            for dst_c_idx in 0..p.c_out {
+            crate::cpu_kernels::par_range(0, p.c_out, num_threads, |dst_c_idx| {
                 let dst_idx = dst_c_idx * l_out;
                 let k_cont = (0..p.c_in)
                     .map(|c_in_idx| k[dst_c_idx * k_s0 + c_in_idx * k_s1 + offset * k_s2])
@@ -1063,10 +1065,14 @@ impl<'a> Map2 for Conv1D<'a> {
                         assert!(k_cont.len() >= p.c_in);
                         let mut d = T::zero();
                         unsafe { T::vec_dot(inp_cont.as_ptr(), k_cont.as_ptr(), &mut d, p.c_in) }
-                        dst[dst_idx] += d
+                        let dst_p = dst.as_ptr();
+                        unsafe {
+                            let ptr = dst_p.add(dst_idx) as *mut T;
+                            *ptr += d
+                        }
                     }
                 }
-            }
+            })
         }
         Ok(dst)
     }