This PR parallelize the conv1d and conv2d ops using the rayon thread pool. They do not use par_iter but rather scoped threads as the amount of work for each task is the same. This gives tighter control on thtread allocation at the expense of using heap allocation.
Benchmarks on my macbook pro (m2pro 2022) from `cpu_benchmarks.rs`:
- conv2d benchmark: 1.29s to 271ms.
- conv1d benchmark: 101ms to 20ms.